+++ rust/worker/src/compactor/mod.rs
@@ -1,2 +1,3 @@
 mod scheduler;
-mod task;
+mod scheduler_policy;
+mod types;
